requirements.txt
# Image Captioning AI Requirements
tensorflow==2.13.0
numpy==1.24.3
pillow>=8.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pandas>=1.3.0
nltk>=3.7

# For evaluation metrics
sacrebleu>=2.0.0
rouge-score>=0.1.2

# Development and utilities
jupyter>=1.0.0
ipykernel>=6.0.0
tqdm>=4.62.0